Updated ctags for windows msysgit again, much more straightforward this time.
Added basic template functionality. Templates are cached in core and
accessed by id or index in the list. This will be changed later to
be accessed by name and id, in the same way that iteminfo works.

By default most information is shown about a template, really only the
'hidden' and 'id' fields are hidden for each field. '-v'erbose will
enable all of these though if you want them.